requirements.txt
Flask==3.0.3
openai==0.28
python-dotenv == 1.0.1